interpreter: Fix DCHECK related to frame pop events
We have a DCHECK that we can only see a retry instruction only on the
second frame after deoptimization. Though we can only pop one frame at a
time, the frame pop event listener can request more such events. So the
expectation that we can see a forced pop only on the second frame isn't
correct.

Test: art/test.py
Bug: 206029744
Change-Id: I258e62861224facb6884d9df1a4cf5406209d731