openjdkjvmti: Remove a warning from redefinition that isn't very useful
We used to warn when there are multiple class definitions in a original
dex file. If we are retransforming and the retransformer doesn't apply
any transformations, then we won't be able to handle it and the warning
was just saying that. However, we do have a check after
retransformations have been applied to check if we have only one class
definition and return an error otherwise. So the warning we are logging
even before retransformation hooks are called is not very useful.

Bug: 259709234
Test: art/test.py
Change-Id: I38b8a20552f887465224da8b9dea926f3b25cc78